added regex to inverter name and MQTT topic (setup.html) beautified serial.html added ticker for wifi loop #515 reverted sunrise / sunset ticker to most recent version

+++ src/utils/llist.h
@@ -11,7 +11,7 @@ struct node_s {
     typedef T dT;
     node_s *pre;
     node_s *nxt;
-    uint32_t id;
+    uint8_t id;
     dT d;
     node_s() : pre(NULL), nxt(NULL), d() {}
     node_s(Args... args) : id(0), pre(NULL), nxt(NULL), d(args...) {}
